refactor(website): extract banner utilities to separate file
- Move fetchEnabledBanners and getDismissedBannerIds to utils.ts
- Add error handling to fetchEnabledBanners with fallback empty array
- Add no-store cache directive for banner data fetching
- Improve code organization and reusability
